+++ src/python/director/cameraview.py
@@ -9,7 +9,7 @@ from director.timercallback import TimerCallback
 from director import vtkNumpy
 from director import objectmodel as om
 import director.vtkAll as vtk
-from director.debugVis import DebugData
+from director.debugpolydata import DebugData
 import vtkRosPython as vtkRos
 
 import PythonQt
