+++ src/python/director/debugpolydata.py
