+++ src/python/director/measurementpanel.py
@@ -7,7 +7,7 @@ from director import objectmodel as om
 from director import vieweventfilter
 from director import visualization as vis
 from director import vtkAll as vtk
-from director.debugVis import DebugData
+from director.debugpolydata import DebugData
 import numpy as np
 
 class MyEventFilter(vieweventfilter.ViewEventFilter):
