+++ src/python/director/pointpicker.py
@@ -2,7 +2,7 @@ import PythonQt
 from PythonQt import QtCore, QtGui
 import director.visualization as vis
 import director.objectmodel as om
-from director.debugVis import DebugData
+from director.debugpolydata import DebugData
 import director.vtkAll as vtk
 from director import callbacks
 import numpy as np
