+++ src/python/director/tasks/robottasks.py
@@ -12,7 +12,7 @@ from director import robotsystem
 from director import transformUtils
 from director import affordanceitems
 from director import vtkNumpy as vnp
-from director.debugVis import DebugData
+from director.debugpolydata import DebugData
 from director import vtkAll as vtk
 #from director import lcmUtils
 import numpy as np
