+++ src/python/tests/testAffordanceItems.py
@@ -5,7 +5,7 @@ from director import affordanceurdf
 from director import objectmodel as om
 from director import visualization as vis
 from director import ioutils
-from director.debugVis import DebugData
+from director.debugpolydata import DebugData
 from director.timercallback import TimerCallback
 from director import segmentation
 from director.uuidutil import newUUID
