+++ src/python/tests/testCameraControl.py
@@ -15,7 +15,7 @@ from director import ioutils
 from director import vtkAll as vtk
 from director import objectmodel as om
 from director import cameracontrolpanel
-from director.debugVis import DebugData
+from director.debugpolydata import DebugData
 
 import numpy as np
